frontend/scripts/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
